ieee1588: Cast a tv_nsec to uint32_t instead of unsigned long

+++ src/qnethernet/drivers/driver_teensy41.cpp
@@ -724,7 +724,7 @@ static struct pbuf* low_level_input(volatile BufferDescriptor* const pBD) {
       p->timestampValid = ((pBD->status & rx_bd_status::kLast) != 0);
       if (p->timestampValid) {
         driver::ieee1588_read_timer(&p->timestamp);
-        if ((unsigned long)p->timestamp.tv_nsec < pBD->timestamp) {
+        if ((uint32_t)p->timestamp.tv_nsec < pBD->timestamp) {
           // The timer has wrapped around
           p->timestamp.tv_sec--;
         }
